Also feed test data in batches
This required a bit of refactoring, too; e.g., generate_batch now
is agnostic to whether it is fed training or test data. It now just
uses an iterator of text window center positions. If it is training
data, this generator is eternal, and reshuffles every X epochs. If
it is testing data, the iterator will run out as usual.